Reactoring DisplayMoviesData, moving file to components
